Downgrade spark-core to 2.7.2
For reasons not yet clear to me, bumping `spark-core` up
to 2.8.0 (the latest version, from Sep 2018) breaks our
server, causing it to throw an exception when we try to
start it up.

Downgrading to 2.7.2 (which is still well ahead of the
old 2.6.0 we were running) seems to fix the problem, so
I'm doing that for now.