Nest `describe`s in add user tests; improve waiting mechanism
This re-nests the add user tests to further reduce duplication.

We also figured out how to eliminate the annoying `setTimeout`
delay which was waiting for the server to respond to the request. We
now use the Protractor expected conditions to wait until the new
user info is present on the page before moving on to the next
expectation. This should make the test more robust to ordering and
other changes to the testing.